+++ src/components/ImageUpload.vue
@@ -65,7 +65,7 @@ watch(() => props.images, () => {
 
         images.value.push(vmImage)
     }
-})
+}, { immediate: true })
 
 const { open, onChange } = useFileDialog({
     accept: 'image/*',
@@ -236,18 +236,20 @@ async function save() {
             </div>
         </div>
         <div class="image-upload__actions">
-            <div>
-                <button :disabled="hasMaxImages" type="button" @click="open()" class="button button--dark">{{ t('imageupload.browse') }}</button>
-            </div>
-            <div>
-                &mdash; {{ t('or') }} &mdash;
-            </div>
-            <div class="image-upload__actions__external">
-                <input v-model="externalImageUrl" class="form-input form-input--small" type="text" :placeholder="t('placeholder.external-image')">
-                <button :disabled="hasMaxImages" type="button" @click="addExternalImage()" class="button button--dark">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="currentColor"><path d="M16.1716 10.9999L10.8076 5.63589L12.2218 4.22168L20 11.9999L12.2218 19.778L10.8076 18.3638L16.1716 12.9999H4V10.9999H16.1716Z"></path></svg>
-                </button>
-            </div>
+            <template v-if="!hasMaxImages">
+                <div>
+                    <button :disabled="hasMaxImages" type="button" @click="open()" class="button button--dark">{{ t('imageupload.browse') }}</button>
+                </div>
+                <div>
+                    &mdash; {{ t('or') }} &mdash;
+                </div>
+                <div class="image-upload__actions__external">
+                    <input v-model="externalImageUrl" class="form-input form-input--small" type="text" :placeholder="t('placeholder.external-image')">
+                    <button :disabled="hasMaxImages" type="button" @click="addExternalImage()" class="button button--dark">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" viewBox="0 0 24 24" fill="currentColor"><path d="M16.1716 10.9999L10.8076 5.63589L12.2218 4.22168L20 11.9999L12.2218 19.778L10.8076 18.3638L16.1716 12.9999H4V10.9999H16.1716Z"></path></svg>
+                    </button>
+                </div>
+            </template>
             <div>
                 {{ t('imageupload.validation', {max: '50MB'}) }} &middot; {{ t('imageupload.status', {current: images.length, max: maxImages}) }}
             </div>